feat(business): BigDecimal money migration (tech-debt #1, slice 23)
Convert business-service monetary values from Float/Double to BigDecimal so
decimal currency is exact (no binary-float rounding drift on totals/dues/reports).

- entities: Sell(sellRate,discount,totalAmount,netAmount,srp), Purchase(totalAmount,
  netAmount), CustomerHistory(paidAmount,dueAmount), Customer(dueAmount),
  Stock(bpurchaseRate,bsellRate,bpurchaseDiscount,bsellDiscount) -> BigDecimal
  with @Column(precision=19, scale=2). Quantities (qty/stock) stay Float.
- matching DTO money fields -> BigDecimal (0.0F defaults -> BigDecimal.ZERO).
- arithmetic rewritten: CustomerService due-amount (add/negate/compareTo),
  SellController revertSell (subtract); StockController default -> ZERO;
  BusinessDashboard sums via doubleValue() + BigDecimal compare.
- gutted dead SellService.createReport (only a commented caller) — its Float
  money math no longer compiles against BigDecimal getters.

Needs build + DB migration #3 (ALTER ... DECIMAL(19,2), in migrations.md) +
headed Cypress. Checked off in docs/tech-debt.md.

+++ microservices/business-service/src/main/java/com/myplus/business_service/dto/CustomerDTO.java
@@ -1,4 +1,5 @@
 package com.myplus.business_service.dto;
+import java.math.BigDecimal;
 
 import java.io.Serializable;
 import java.time.LocalDate;
@@ -41,9 +42,9 @@ public class CustomerDTO implements Serializable {
 	private String address;
 
 	// @ValidateEmpty(message = "Paid amount is required")
-	private Float paidAmount;
+	private BigDecimal paidAmount;
 
-    private Float dueAmount;
+    private BigDecimal dueAmount;
 
     private LocalDate dueDate;
